+++ source/extensions/omni.isaac.surface_gripper/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.3.2] - 2022-12-01
+### Fixed
+- CreateSurfaceGripper command documentation update
+
 
 ## [0.3.1] - 2022-12-01
 ### Fixed